model loading, re-org, added comments
